test: typeorm composite foreign key

+++ backend/node/app/src/util/entity/game.log.entity.ts
@@ -21,7 +21,7 @@ export default class GameLogs {
   @Column('integer')
   playerBId: number;
 
-  @Column('tinyint')
+  @Column('smallint')
   result: number;
 
   @CreateDateColumn()
